Add a docker controller for tasks concerning docker
